requirements.txt
Django>=4.2,<5.0
Pillow>=10.0
numpy>=1.24
opencv-python-headless>=4.8
easyocr>=1.7
Jinja2>=3.1
python-dotenv>=1.0
